+++ packages/gasket-plugin-intl/lib/build-manifest.js
@@ -3,7 +3,6 @@
 const path = require('path');
 const fs = require('fs').promises;
 const { promisify } = require('util');
-const loaderUtils = require('loader-utils');
 const { getIntlConfig } = require('./configure');
 
 // TODO: Need to review for native promise usage
@@ -15,20 +14,25 @@ const debug = require('debug')('gasket:plugin:intl:buildManifest');
  * Constructs a manifest of locale file paths and settings which can be loaded
  * or bundled by the client. Locale paths have content hashes associated with
  * them which can be used for cache busting.
- * @param {import("@gasket/core").Gasket} gasket - Gasket API
+ * @param {import('@gasket/core').Gasket} gasket - Gasket API
  * @async
  */
 module.exports = async function buildManifest(gasket) {
   const { logger } = gasket;
-  const { localesDir, manifestFilename } = getIntlConfig(gasket);
-  const tgtFile = path.join(localesDir, manifestFilename);
-  const { basePath, defaultPath, defaultLocale, locales, localesMap } =
-    getIntlConfig(gasket);
+  const { localesDir, managerFilename } = getIntlConfig(gasket);
+  const tgtFile = path.join(gasket.config.root, ...managerFilename.split('/'));
+  const {
+    defaultLocaleFilePath,
+    staticLocaleFilePaths,
+    defaultLocale,
+    locales,
+    localesMap
+  } = getIntlConfig(gasket);
 
   // find all the .json files except the manifest
   debug(`Building manifest ${tgtFile} from JSON files in ${localesDir}`);
   const files = (await glob('**/*.json', { cwd: localesDir })).filter(
-    (f) => f !== manifestFilename
+    (f) => f !== managerFilename
   );
 
   if (!files.length) {
@@ -36,33 +40,47 @@ module.exports = async function buildManifest(gasket) {
   }
 
   // generate a content hash for each file
-  const paths = (
-    await Promise.all(
-      files.map(async (file) => {
-        const buffer = await fs.readFile(path.join(localesDir, file));
-        const hash = loaderUtils.getHashDigest(buffer, 'md5', 'hex', 7);
-        return { [path.basename(localesDir) + '/' + file]: hash };
-      })
-    )
+  const imports = (
+    files.map((file) => {
+      const importName = ['.', path.basename(localesDir), file].join('/');
+      const keyName = importName
+        .replace(/\.json$/, '')
+        .replace('./', '');
+      return { [keyName]: `%() => import('${importName}')%` };
+    })
   ).reduce((a, c) => ({ ...a, ...c }), {});
 
   /**
    * Locale settings and known locale file paths
    */
   const manifest = {
-    basePath,
-    defaultPath,
-    defaultLocale,
-    locales,
-    localesMap,
-    paths
+    '%defaultLocaleFilePath%': defaultLocaleFilePath,
+    '%staticLocaleFilePaths%': staticLocaleFilePaths,
+    '%defaultLocale%': defaultLocale,
+    '%locales%': locales,
+    '%localesMap%': localesMap,
+    '%imports%': imports
   };
 
+  let manifestStr = JSON.stringify(manifest, null, 2);
+  manifestStr = manifestStr.replace(/("%|%")/g, '');
+  manifestStr = manifestStr.replace(/"/g, `'`);
+
+  const content = `/* -- GENERATED FILE - DO NOT EDIT -- */
+import { makeIntlManager } from '@gasket/helper-intl';
+
+const manifest = ${manifestStr};
+
+export default makeIntlManager(manifest);
+`;
+
+  const tgtRelative = path.relative(gasket.config.root, tgtFile);
+
   try {
-    await fs.writeFile(tgtFile, JSON.stringify(manifest), 'utf-8');
-    logger.info('build:locales: Wrote locales manifest.');
+    await fs.writeFile(tgtFile, content, 'utf-8');
+    logger.info(`build:locales: Wrote intl manager ${tgtRelative}.`);
   } catch (err) {
-    logger.error('build:locales: Unable to write locales manifest.');
+    logger.error(`build:locales: Unable to write intl manager (${tgtRelative}).`);
     throw err;
   }
 };