+++ packages/gasket-plugin-intl/lib/configure.js
@@ -9,20 +9,13 @@ const moduleDefaults = {
   excludes: ['cacache', 'yargs', 'axe-core']
 };
 
-const isDefined = (o) => typeof o !== 'undefined';
-
 /**
  * Shortcut to get the gasket.config.intl object
  * @param {import("@gasket/core").Gasket} gasket - Gasket API
  * @returns {import('./index').IntlConfig} intl config
  */
 function getIntlConfig(gasket) {
-  const { intl = {} } = gasket.config || {};
-
-  // handling default here is necessary for metadata which runs before configure
-  // hook
-  intl.localesDir = intl.localesDir || path.join('public', 'locales');
-  return intl;
+  return gasket.config.intl;
 }
 
 /**
@@ -35,27 +28,36 @@ module.exports = function configure(gasket, config) {
   // @ts-ignore - temp fix until we can get types for gasket-plugin-intl
   const intlConfig = { ...getIntlConfig({ config }) };
 
-  const { nextConfig = {} } = config;
-
   // get user defined config and apply defaults
   const {
-    defaultPath = '/locales',
-    defaultLocale = 'en',
+    locales,
     localesMap = {},
-    localesDir,
-    manifestFilename = 'locales-manifest.json',
-    preloadLocales = false
+    defaultLocaleFilePath = 'locales',
+    localesDir = 'locales',
+    managerFilename = 'intl.js'
+  } = intlConfig;
+
+  let {
+    defaultLocale,
+    staticLocaleFilePaths
   } = intlConfig;
 
   const fullLocalesDir = path.join(root, localesDir);
 
-  const basePath = [
-    intlConfig.basePath,
-    nextConfig.assetPrefix,
-    nextConfig.basePath,
-    config.basePath,
-    ''
-  ].find(isDefined);
+  if (!locales || !locales.length) {
+    throw new Error('Gasket config required for intl.locales');
+  }
+
+  if (!defaultLocale) {
+    defaultLocale = locales[0];
+    gasket.logger.debug(`intl.defaultLocale not configured, defaulting to first intl.locales (${defaultLocale})`);
+  }
+
+  if (!staticLocaleFilePaths) {
+    staticLocaleFilePaths = [defaultLocaleFilePath];
+    gasket.logger.debug(`intl.staticLocaleFilePaths not configured, defaulting to ([${staticLocaleFilePaths.join(', ')}])`);
+  }
+
 
   let { modules = false } = intlConfig;
   if (modules && !Array.isArray(modules)) {
@@ -63,27 +65,15 @@ module.exports = function configure(gasket, config) {
       modules === true ? moduleDefaults : { ...moduleDefaults, ...modules };
   }
 
-  /* eslint-disable no-process-env */
-
-  // Allows @gasket/react-intl/next to perform server side loading
-  process.env.GASKET_INTL_LOCALES_DIR = fullLocalesDir;
-
-  // Allows @gasket/react-intl to access manifest, and is bundled for browser
-  process.env.GASKET_INTL_MANIFEST_FILE = path.join(
-    fullLocalesDir,
-    manifestFilename
-  );
-  /* eslint-enable no-process-env */
-
   const normalizedIntlConfig = {
     ...intlConfig,
-    basePath,
-    defaultPath,
     defaultLocale,
+    locales,
     localesMap,
+    defaultLocaleFilePath,
+    staticLocaleFilePaths,
     localesDir: fullLocalesDir,
-    manifestFilename,
-    preloadLocales,
+    managerFilename,
     modules
   };
 
@@ -91,14 +81,12 @@ module.exports = function configure(gasket, config) {
 
   /**
    * @typedef {object} IntlConfig
-   * @property {string} basePath - Base URL where locale files are served
-   * @property {string} defaultPath - Path to endpoint with JSON files
+   * @property {string} defaultLocaleFilePath - Path to endpoint with JSON files
    * @property {string} defaultLocale - Locale to fallback to when loading files
    * @property {object} localesMap - Mapping of locales to share files
    * @property {string} localesDir - Path to on-disk directory where locale
    * files exists
-   * @property {string} manifestFilename - Name of the manifest file
-   * @property {boolean} preloadLocales - Preloads locale files if set to true
+   * @property {string} managerFilename - Name of the manager file
    * @property {object} modules - Enable locale files collation from node
    * modules
    * @property {string} modules.localesDir - Lookup dir for module files