+++ packages/gasket-react-intl/jest.config.js
@@ -1,5 +1,5 @@
-module.exports = {
-  collectCoverageFrom: ['lib/**/*.js'],
+export default {
+  collectCoverageFrom: ['src/**/*.js'],
   setupFiles: ['<rootDir>/test/setup.js'],
   testEnvironment: 'jsdom'
 };